+++ roles/setup_base_cli/files/neovim/.config/nvim/lua/system-forger/init.lua
@@ -1,0 +1,3 @@
+-- THIS FILE IS MANAGED BY ANSIBLE; ANY MANUAL CHANGES WILL BE LOST
+
+require("system-forger.lazy")