+++ CHANGELOG.md
@@ -49,6 +49,20 @@ and this project adheres to [Semantic Versioning].
     integration. In case `fd` gets removed it will **NOT** revert to using
     `find`. `fd` and `find` are not CLI or partially CLI compatible.
 
+  - **neovim:** Install `neovim` with basic configuration:
+    - Highlighting current line
+    - Relative line numbers
+    - Smart case text search
+    - 80 char marker line
+    - Use spaces instead of tabs
+    - Ergonomic splits management: split horizontally, split vertically, close,
+      make splits equal size.
+    - Ergonomic tabs management: create new tab, close tab, go to next tab, go
+      to previous tab.
+    - Move selected lines up or down
+    - Fuzzy finder using `telescope` plugin. `lazy.nvim` is used as plugin
+      manager.
+
   - **btop:** Modern system monitor for the CLI. Provided monitoring for disk
     usage, CPU and GPU loads, current processes, network upload and download
     speed overview.